Update test11_no_ultrasonic.ino
Added variables so that c = equals the case number the younity bot is currently on (ex. when case 4 is active, c = 4). This will allow for us to code in the autonomous cases: 

while (c = case #) { 
do this 
}

This will allow for the autonomous programs to run ONLY when their cases are being called.